+++ Bank.DAL/Views/CreateTransactionRequestView.cs
@@ -1,2 +1,2 @@
 ﻿namespace Bank.DAL.Views;
-    public record CreateTransactionRequestView(int Sum, string CardNumber);
+    public record CreateTransactionRequestView(double Sum, string CardNumber);
