+++ Market/Transaction.cs
@@ -2,7 +2,7 @@
 {
     public class Transaction
     {
-        public int Sum { get; init; }
+        public double Sum { get; init; }
         public string CardNumber { get; init; }
     }
 }
